move beat manifest
