+++ frontend/src/app/domain/ports/beat-metadata.ts
@@ -1,0 +1,6 @@
+export type BeatMetadata = {
+  readonly genre: string;
+  readonly label: string;
+  readonly filename: string;
+  readonly bpm?: number;
+};
