+++ frontend/src/app/domain/ports/i-manage-beats.ts
@@ -1,7 +1,7 @@
 import { Effect } from "effect";
 import { Beat } from "../beat";
 import { Track } from "../track";
-import {BeatMetadata} from "../../../assets/beats/beats-manifest";
+import {BeatMetadata} from "./beat-metadata";
 
 export default interface IManageBeats {
   readonly getBeatByFileName: (filename: string) => Effect.Effect<Beat, Error>
